validation: pktio: fix start_stop test
Test of Rx on a stopped pktio was broken due to a missing:
+                               pkt = odp_packet_from_event(ev);

pkt value was the last return for odp_packet_alloc so
in the setup ODP_PACKET_INVALID which made it work. However changing
the size of the pool to > 100 changed pkt value so it points to
a valid packet and tes would fail.

However setting pkt to its right value from an event break the test for
loopback, because the packets sent obliouvys come back and are returned
by the calls to odp_schedule().
Thus I have disable this half of the start stop test if there is only
 one pktio provided.

Signed-off-by: Nicolas Morey-Chaisemartin <nmorey@kalray.eu>